feat(sim): add CPU-only simulated engine
