+++ pegainfer-core/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+pegainfer-engine = { workspace = true }
 pegainfer-kernels = { workspace = true }
 anyhow = { workspace = true }
 colored = { workspace = true }
