+++ pegainfer-engine/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "pegainfer-engine"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+tokio = { workspace = true, features = ["sync"] }
+
+[lints]
+workspace = true
