+++ pegainfer-engine/src/engine.rs
@@ -1,0 +1,105 @@
+use std::path::PathBuf;
+
+use tokio::sync::mpsc;
+
+use crate::sampler::SamplingParams;
+
+#[derive(Clone, Debug)]
+pub struct EngineLoadOptions {
+    pub enable_cuda_graph: bool,
+    pub enable_prefill_profile: bool,
+    pub device_ordinals: Vec<usize>,
+    pub seed: u64,
+}
+
+impl Default for EngineLoadOptions {
+    fn default() -> Self {
+        Self {
+            enable_cuda_graph: true,
+            enable_prefill_profile: false,
+            device_ordinals: vec![0],
+            seed: 42,
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct ModelInfo {
+    pub id: &'static str,
+    pub display_name: String,
+    pub model_path: PathBuf,
+    pub max_model_len: Option<u32>,
+}
+
+#[derive(Clone, Debug)]
+pub struct TokenLogprob {
+    pub logprob: f32,
+    pub top_logprobs: Vec<(u32, f32)>,
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum FinishReason {
+    Length,
+    Stop,
+    Error,
+}
+
+pub struct GenerateRequest {
+    pub request_id: Option<String>,
+    pub queued_at_unix_s: Option<f64>,
+    pub prompt_tokens: Vec<u32>,
+    pub params: SamplingParams,
+    pub max_tokens: usize,
+    pub token_tx: mpsc::UnboundedSender<TokenEvent>,
+    pub logprobs: usize,
+    pub echo: bool,
+}
+
+pub enum TokenEvent {
+    Scheduled {
+        queued_at_unix_s: f64,
+        scheduled_at_unix_s: f64,
+        prompt_tokens: usize,
+    },
+    Token {
+        id: u32,
+        logprob: Option<TokenLogprob>,
+    },
+    PromptTokens {
+        ids: Vec<u32>,
+        logprobs: Vec<Option<TokenLogprob>>,
+    },
+    Finished {
+        finish_reason: FinishReason,
+        prompt_tokens: usize,
+        completion_tokens: usize,
+    },
+    Error {
+        message: String,
+        prompt_tokens: usize,
+        completion_tokens: usize,
+    },
+    Rejected {
+        message: String,
+        prompt_tokens: usize,
+        completion_tokens: usize,
+    },
+}
+
+#[derive(Clone)]
+pub struct EngineHandle {
+    submit_tx: mpsc::UnboundedSender<GenerateRequest>,
+}
+
+impl EngineHandle {
+    pub fn new(submit_tx: mpsc::UnboundedSender<GenerateRequest>) -> Self {
+        Self { submit_tx }
+    }
+
+    pub fn submit(
+        &self,
+        req: GenerateRequest,
+    ) -> std::result::Result<(), mpsc::error::SendError<GenerateRequest>> {
+        self.submit_tx.send(req)
+    }
+}
