+++ pegainfer-engine/src/lib.rs
@@ -1,0 +1,4 @@
+//! Lightweight engine request/event contract shared by real and simulated engines.
+
+pub mod engine;
+pub mod sampler;
