+++ pegainfer-sim/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "pegainfer-sim"
+version = "0.1.0"
+edition = "2024"
+
+[[bin]]
+name = "pegainfer-sim"
+path = "src/main.rs"
+
+[dependencies]
+anyhow = { workspace = true }
+clap = { workspace = true }
+pegainfer-engine = { workspace = true }
+pegainfer-vllm-frontend = { workspace = true }
+tokio = { workspace = true, features = ["full"] }
+
+[lints]
+workspace = true
